+++ src/main/java/org/ezcode/codetest/infrastructure/notification/event/NotificationSavedEvent.java
@@ -1,0 +1,12 @@
+package org.ezcode.codetest.infrastructure.notification.event;
+
+import org.ezcode.codetest.infrastructure.notification.dto.NotificationResponse;
+
+public record NotificationSavedEvent(
+
+	String principalName,
+
+	NotificationResponse response
+
+) {
+}
